+++ codegen/src/rust.rs
@@ -1077,13 +1077,13 @@ impl<'a> RustGenerator<'a> {
         if self.options.introspection {
             if let Some(feature) = self.rust_options.introspection_if {
                 genln!(self, "        #[cfg(feature = \"{feature}\")]");
-                genln!(self, "        let info = aldrin::core::ServiceInfo::new(Self::VERSION).set_type_id(<Self as aldrin::core::introspection::Introspectable>::type_id());");
+                genln!(self, "        let info = aldrin::low_level::ServiceInfo::new(Self::VERSION).set_type_id(<Self as aldrin::core::introspection::Introspectable>::type_id());");
 
                 genln!(self, "        #[cfg(not(feature = \"{feature}\"))]");
-                genln!(self, "        let info = aldrin::core::ServiceInfo::new(Self::VERSION);");
+                genln!(self, "        let info = aldrin::low_level::ServiceInfo::new(Self::VERSION);");
             }
         } else {
-            genln!(self, "        let info = aldrin::core::ServiceInfo::new(Self::VERSION);");
+            genln!(self, "        let info = aldrin::low_level::ServiceInfo::new(Self::VERSION);");
         }
 
         genln!(self, "        let inner = object.create_service(Self::UUID, info).await?;");